requirements.txt
numpy~=1.23.4
pytesseract~=0.3.10
keyboard~=0.13.5
Pillow~=9.3.0